+++ app/forms/manage.py
@@ -23,8 +23,16 @@ class DeviceForm(FlaskForm):
     def __init__(self, is_developer, *args, **kwargs):
         super(DeviceForm, self).__init__(*args, **kwargs)
         if is_developer:
-            self.device_type.choices = [('', '选择设备类型')] + [(str(device_type.id), device_type.name) for device_type in DeviceType.query.order_by(DeviceType.id.asc()).all()]
+            self.device_type.choices = [('', '选择设备类型')] + \
+                [(str(device_type.id), device_type.name) \
+                for device_type in DeviceType.query.order_by(DeviceType.id.asc()).all()]
         else:
-            self.device_type.choices = [('', '选择设备类型')] + [(str(device_type.id), device_type.name) for device_type in DeviceType.query.filter(DeviceType.name != 'Server').order_by(DeviceType.id.asc()).all()]
-        self.room.choices = [('', '选择所属场地')] + [('0', '无')] + [(str(room.id), room.name) for room in Room.query.order_by(Room.id.asc()).all()]
-        self.lesson_types.choices = [('', '选择授权课程类型')] + [(str(lesson_type.id), lesson_type.name) for lesson_type in LessonType.query.all()]
+            self.device_type.choices = [('', '选择设备类型')] + \
+                [(str(device_type.id), device_type.name) for device_type in DeviceType.query\
+                    .filter(DeviceType.name != 'Server')\
+                    .order_by(DeviceType.id.asc())\
+                    .all()]
+        self.room.choices = [('', '选择所属场地')] + [('0', '无')] + \
+            [(str(room.id), room.name) for room in Room.query.order_by(Room.id.asc()).all()]
+        self.lesson_types.choices = [('', '选择授权课程类型')] + \
+            [(str(lesson_type.id), lesson_type.name) for lesson_type in LessonType.query.all()]